move run-fonts to docx/run and add a test
